Changing styling of buttons and search
Removed redundant code and fixed errors. Added styling to the search bar and next/back buttons. Added additional functionality so that the user is taken to the top of the page when they click a button or pokemon card.

Going to work on fixing the unique keys bug and hiding the next/back buttons properly.